requirements.txt
googletrans==3.1.0a0
keybert==0.7.0
openai==0.25.0
pydub==0.25.1
python-dotenv==0.21.1
streamlit==1.17.0
sumy==0.11.0
transformers==4.25.1
